Update to staff page
FIXED - A non-existent bug with mailer.php not sending email, (issue was
found to be with godaddy's hosting.)
CHANGED - Size of navigation font from 2em to 1.8em to make the width of
the navigation the same as the page width.
CHANGED - Added the Facebook link to the get b online page.